Add files via upload
CloudBurst, an application using Matlab that analyzes the weather forecasts and predicts the amount of rainfall in the near future. Past history data of the weather forecasts are fed and by using statistical tools and concepts like multiple regression and exponential smoothing method, a reliable future rainfall prediction along with graph plots and graphs comparisons are generated.